# api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: contact@josehernandez.dev
    name: Smidgen
    url: https://example.com/
  description: API for interacting with Smidgen.
  license:
    name: " GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007"
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  title: Smidgen
  version: 0.0.1
servers:
- url: /api/v1
paths:
...
  /equipment/:
    get:
      description: Get all equipments stored in the database.
      operationId: get_equipment_equipment__get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/equipment'
                title: Response Get Equipment Equipment  Get
                type: array
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
      summary: Get equipments
      tags:
      - equipment
    post:
      description: Create a new equipment in the database.
      operationId: add_equipment_equipment__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/equipment'
        required: true
      responses:
        "202":
          content:
            application/json:
              schema: {}
          description: The data received was accepted and is quered for addition to
            the database.
        "200":
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create equipment
      tags:
      - equipment
  /equipment/{equipment_id}:
    delete:
      description: Delete the specified equipment from the database.
      operationId: delete_equipment_equipment__equipment_id__delete
      parameters:
      - explode: false
        in: path
        name: equipment_id
        required: true
        schema:
          title: Equipment Id
          type: integer
        style: simple
      responses:
        "202":
          content:
            application/json:
              schema: {}
          description: The data received was accepted and is quered for addition to
            the database.
        "200":
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete equipment
      tags:
      - equipment
    get:
      description: Get the specified equipment stored in the database.
      operationId: get_equipments_equipment__equipment_id__get
      parameters:
      - explode: false
        in: path
        name: equipment_id
        required: true
        schema:
          title: Equipment Id
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/equipment'
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get equipment
      tags:
      - equipment
    put:
      description: Update the specified equipment in the database.
      operationId: update_equipment_equipment__equipment_id__put
      parameters:
      - explode: false
        in: path
        name: equipment_id
        required: true
        schema:
          title: Equipment Id
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/equipment'
        required: true
      responses:
        "202":
          content:
            application/json:
              schema: {}
          description: The data received was accepted and is quered for addition to
            the database.
        "200":
          description: The data was found and has been returned.
        "404":
          description: The data requested was not found in the database.
        "401":
          description: You are unauthorized to view this resource.
        "403":
          description: Access to this page is forbidden. Please reference the API
            documentation for more information.
        "500":
          description: An unexpected error has occured.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update equipment
      tags:
      - equipment
...
components:
  schemas:
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - Location_inner
          - Location_inner
          type: type
        - msg: msg
          loc:
          - Location_inner
          - Location_inner
          type: type
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      example:
        msg: msg
        loc:
        - Location_inner
        - Location_inner
        type: type
      properties:
        loc:
          items:
            $ref: '#/components/schemas/Location_inner'
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
    equipment:
      example:
        business_unit_id: 6
        description: description
        model: model
        last_inventoried: 2000-01-23T04:56:07.000+00:00
        date_received: 2000-01-23T04:56:07.000+00:00
        equipment_id: 0
        manufacturer: manufacturer
      properties:
        equipment_id:
          title: Equipment Id
          type: integer
        business_unit_id:
          title: Business Unit Id
          type: integer
        manufacturer:
          title: Manufacturer
          type: string
        model:
          title: Model
          type: string
        description:
          title: Description
          type: string
        date_received:
          format: date-time
          title: Date Received
          type: string
        last_inventoried:
          format: date-time
          title: Last Inventoried
          type: string
      required:
      - business_unit_id
      - date_received
      - description
      - equipment_id
      - last_inventoried
      - manufacturer
      - model
      title: equipment
      type: object
...
    Location_inner:
      anyOf:
      - type: string
      - type: integer